refactor(rpc): Remove all suffix of RPC type as their is no difference between 005 and 006
